docs(storybook): more storybook work
