+++ packages/badge/src/components/Badge.stories.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
 import { storiesOf, StoryDecorator } from '@storybook/react';
 
-import Wrapper from '@storybook-utils/components/Wrapper';
+import Wrapper from '../../../../.storybook/components/Wrapper';
 
 import Badge from './Badge';
 
-import readme from '../../README.md';
+const readme = require('../../README.md');
 
 const StoryWrapper: StoryDecorator = storyFn => <Wrapper>{storyFn()}</Wrapper>;
