+++ packages/modal/tsconfig.json
@@ -11,5 +11,11 @@
     { "path": "../theme" },
     { "path": "../common" }
   ],
-  "exclude": ["**/__stories__", "**/__tests__", "**/__mocks__", "**/lib"]
+  "exclude": [
+    "**/*.stories.tsx",
+    "**/*.story.tsx",
+    "**/__tests__",
+    "**/__mocks__",
+    "**/lib"
+  ]
 }
