+++ packages/pagination/tsconfig.json
@@ -7,5 +7,11 @@
     "outDir": "tsc-out"
   },
   "references": [{ "path": "../button" }, { "path": "../theme" }],
-  "exclude": ["**/__stories__", "**/__tests__", "**/__mocks__", "**/lib"]
+  "exclude": [
+    "**/*.stories.tsx",
+    "**/*.story.tsx",
+    "**/__tests__",
+    "**/__mocks__",
+    "**/lib"
+  ]
 }
