+++ packages/theme/tsconfig.json
@@ -6,5 +6,11 @@
     "rootDir": "src",
     "outDir": "tsc-out"
   },
-  "exclude": ["**/__stories__", "**/__tests__", "**/__mocks__", "**/lib"]
+  "exclude": [
+    "**/*.stories.tsx",
+    "**/*.story.tsx",
+    "**/__tests__",
+    "**/__mocks__",
+    "**/lib"
+  ]
 }
